# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.CreateNewsletterStruct:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterInviteStruct:
    properties:
      key:
        type: string
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterMessagesStruct:
    properties:
      before_id:
        type: integer
      count:
        type: integer
      jid:
        $ref: '#/definitions/go_mau_fi_whatsmeow_types.JID'
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterStruct:
    properties:
      jid:
        $ref: '#/definitions/go_mau_fi_whatsmeow_types.JID'
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_poll_model.PollResults:
    properties:
      optionCounts:
        additionalProperties:
          type: integer
        description: hash -> count
        type: object
      pollChatJid:
        type: string
      pollMessageId:
        type: string
      totalVotes:
        type: integer
      voters:
        items:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_poll_model.VoterInfo'
        type: array
      votes:
        items:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_poll_model.PollVote'
        type: array
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_poll_model.PollVote:
    properties:
      companyId:
        type: string
      id:
        type: string
      instanceId:
        type: string
      pollChatJid:
        type: string
      pollMessageId:
        type: string
      receivedAt:
        type: string
      selectedOptions:
        description: SHA-256 hashes
        items:
          type: string
        type: array
      voteMessageId:
        type: string
      votedAt:
        type: string
      voterJid:
        type: string
      voterName:
        type: string
      voterPhone:
        type: string
    type: object
  github_com_WebAPPWago_webapp-wago_pkg_poll_model.VoterInfo:
    properties:
      jid:
        type: string
      name:
        type: string
      phone:
        type: string
      selectedOptions:
        items:
          type: string
        type: array
      votedAt:
        type: string
    type: object
...
  go_mau_fi_whatsmeow_types.JID:
    properties:
      device:
        format: int32
        type: integer
      integrator:
        format: int32
        type: integer
      rawAgent:
        format: int32
        type: integer
      server:
        type: string
      user:
        type: string
    type: object
...
info:
  contact: {}
  description: WebAPP-Wago - whatsmeow
  title: WebAPP-Wago
  version: "1.0"
paths:
...
  /newsletter/create:
    post:
      consumes:
      - application/json
      description: Create newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.CreateNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create newsletter
      tags:
      - Newsletter
  /newsletter/info:
    post:
      consumes:
      - application/json
      description: Get newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter
      tags:
      - Newsletter
  /newsletter/link:
    post:
      consumes:
      - application/json
      description: Get newsletter invite
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterInviteStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter invite
      tags:
      - Newsletter
  /newsletter/list:
    get:
      consumes:
      - application/json
      description: List newsletters
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List newsletters
      tags:
      - Newsletter
  /newsletter/messages:
    post:
      consumes:
      - application/json
      description: Get newsletter messages
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterMessagesStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get newsletter messages
      tags:
      - Newsletter
  /newsletter/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe newsletter
      parameters:
      - description: Newsletter data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_newsletter_service.GetNewsletterStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Subscribe newsletter
      tags:
      - Newsletter
  /polls/{pollMessageId}/results:
    get:
      consumes:
      - application/json
      description: Retorna todos os votos de uma enquete específica
      parameters:
      - description: ID da mensagem da enquete
        in: path
        name: pollMessageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_WebAPPWago_webapp-wago_pkg_poll_model.PollResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get poll results
      tags:
      - Polls
...
swagger: "2.0"